+++ CHANGES.md
@@ -2,3 +2,4 @@
 
 - Restructured the HTML output: each term in its own subsection instead of into a table (this gives more space for the content)
 - Added the possibility to add one or more examples into the YML file
+- The YAML content is checked with a built-in JSON Schema before processing further